@@ -1,52 +1,68 @@
 {
   "name": "weather-app",
   "private": true,
+  "type": "module",
   "scripts": {
     "dev": "cross-env NODE_ENV='development' NODE_OPTIONS='--inspect' next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint"
+    "lint": "next lint --fix",
+    "test": "NODE_ENV='test' vitest --coverage --reporter=verbose",
+    "ci": "NOD_ENV='prod' vitest --coverage",
+    "audit": "better-npm-audit audit"
   },
   "dependencies": {
     "@emotion/cache": "latest",
     "@emotion/react": "^11.11.4",
     "@emotion/server": "latest",
     "@emotion/styled": "^11.11.5",
-    "@mdi/js": "^6.9.96",
+    "@faker-js/faker": "^8.4.1",
+    "@mdi/js": "^7.4.47",
     "@mdi/react": "^1.6.1",
     "@mui/icons-material": "^5.15.20",
     "@mui/material": "^5.15.20",
     "@mui/material-nextjs": "^5.15.11",
     "@mui/system": "^5.15.20",
+    "@next/bundle-analyzer": "^14.2.4",
     "@tanstack/react-query": "^5.45.1",
-    "@tanstack/react-query-devtools": "5.40.1",
+    "@tanstack/react-query-devtools": "5.45.1",
     "@types/luxon": "^3.4.2",
-    "@types/react": "^18.3.3",
     "@types/request-ip": "^0.0.41",
     "@vercel/analytics": "^1.3.1",
     "@vercel/speed-insights": "^1.0.12",
     "luxon": "^3.4.4",
     "next": "^14.2.4",
     "npm": "^10.8.1",
-    "prop-types": "^15.8.1",
     "react": "latest",
     "react-dom": "latest",
     "recharts": "2.13.0-alpha.4",
     "request-ip": "^3.3.0",
     "three-dots": "^0.3.2",
     "throttle-debounce": "^5.0.0",
-    "uuid": "^9.0.1",
     "zod": "^3.23.8",
     "zustand": "^4.5.2"
   },
   "devDependencies": {
-    "@types/node": "20.12.13",
-    "@typescript-eslint/eslint-plugin": "^7.11.0",
-    "@typescript-eslint/parser": "^7.11.0",
+    "@testing-library/dom": "^10.1.0",
+    "@testing-library/jest-dom": "^6.4.6",
+    "@testing-library/react": "^16.0.0",
+    "@testing-library/user-event": "^14.5.2",
+    "@types/node": "20.14.2",
+    "@types/react": "^18.3.3",
+    "@types/react-dom": "^18.3.0",
+    "@typescript-eslint/eslint-plugin": "^7.13.0",
+    "@typescript-eslint/parser": "^7.13.0",
+    "@vitejs/plugin-react": "^4.3.1",
+    "@vitest/coverage-v8": "^1.6.0",
     "cross-env": "^7.0.3",
-    "eslint": "^8.*.*",
-    "eslint-config-next": "14.2.3",
-    "typescript": "^5.4.5"
+    "eslint": "^8.56.0",
+    "eslint-config-next": "14.2.4",
+    "eslint-plugin-unused-imports": "^4.0.0",
+    "jsdom": "^24.1.0",
+    "msw": "^2.3.1",
+    "typescript": "^5.4.5",
+    "vite-tsconfig-paths": "^4.3.2",
+    "vitest": "^1.6.0"
   },
   "packageManager": "pnpm@9.1.1+sha512.14e915759c11f77eac07faba4d019c193ec8637229e62ec99eefb7cf3c3b75c64447882b7c485142451ee3a6b408059cdfb7b7fa0341b975f12d0f7629c71195"
 }
