@@ -1,5 +1,10 @@
 // @ts-check
- 
+import withBundleAnalyzer from '@next/bundle-analyzer';
+
+const bundleAnalyzer = withBundleAnalyzer({
+	enabled: process.env.ANALYZE === 'true',
+})
+
 /** @type {import('next').NextConfig} */
 const nextConfig = {
     logging: {
@@ -9,4 +14,4 @@ const nextConfig = {
     },
 }
    
-module.exports = nextConfig
+export default bundleAnalyzer(nextConfig);
