@@ -1,7 +1,7 @@
 import { NextRequest } from "next/server";
 import { getNearbyLocationDetails } from "../../../../src/services/geolocation.service";
-import { extractQueryParams } from "../../../../src/utils/url";
-import { coordsSchema } from '../../../../src/schemas/coords.schema';
+import { extractQueryParams, handleNextResponseError } from "../../../../src/utils/next-request-helper";
+import { coordsSchema } from '../../../../src/features/weaget/coords.schema';
 
 // Generates a list of autocomplete queries given a string
 export async function GET(req: NextRequest) {
@@ -18,9 +18,6 @@ export async function GET(req: NextRequest) {
         return Response.json(formattedLocations);
     }
     catch (err) {
-        console.error(err);
-        const errorId = crypto.randomUUID();
-        const errorMessage = `Failed to retrieve nearby locations.`;
-        return Response.json({id: errorId, message: errorMessage}, {status: 500 });
+        return handleNextResponseError(err, 'Failed to retrieve nearby locations.');
     }
 }
