@@ -1,8 +1,8 @@
 import { NextRequest } from "next/server";
-import { coordsSchema } from '../../../src/schemas/coords.schema';
+import apicnPollutionSchema from "../../../src/features/apicn-pollution/pollution.schema";
+import { coordsSchema } from '../../../src/features/weaget/coords.schema';
 import { getPollutionByCoord } from "../../../src/services/pollution.service";
-import { extractQueryParams } from "../../../src/utils/url";
-import apicnPollutionSchema from "../../../src/schemas/apicn/pollution.schema";
+import { extractQueryParams, handleNextResponseError } from "../../../src/utils/next-request-helper";
 
 export async function GET(req: NextRequest) {
     const queryParams = extractQueryParams(`${req.url}`);
@@ -14,9 +14,6 @@ export async function GET(req: NextRequest) {
         return Response.json(response);
     }
     catch (err) {
-        console.error(err);
-        const errorId = crypto.randomUUID();
-        const errorMessage = `Failed to retrieve pollution data.`;
-        return Response.json({id: errorId, message: errorMessage}, {status: 500 });
+        return handleNextResponseError(err, 'Failed to retrieve pollution data.');
     }
 }
